Update main.py
First change to note is I changed the name from "main.py" to "discord-bot.py". This was done so that the general function of files are easier to distinguish at first glance.

Within my code I initially decided to create an instance of a Client object, the Bot's connection and way of interacting with the Discord Client. However through further research found that I would rather create an instance of a Bot over a Client.

In this particular Python Library, discord.py, a Bot is a subclass of a Client. In other words, the Bot class is an extension of a the Client class. I would have access to all of the Client methods and attributes, in addition to the attributes and methods that were created for the Bot class.

Long story made short, I decided to use a Bot instance over a client instance to gain access to the command functionality. With this functionality I can set a command prefix, and define commands that would follow the command prefix.

Whereas if I continued to move forward with a Client instance, I would have had to register on the event of a message being sent into a channel, the bot would check if the message started with a particular string.

Using the Bot Client, also helped to clean up my code by allowing me to reduce the amount of if-statements I had and making the code easier to read.